fixes to enable compilation with ICC
The primary changes are avoiding pointer arithmetic on void * types, which is
evidently a GCCism: instead a byte-sized pointer type should be used.

Other changes include workarounds to avoid uninitialised variable warnings,
and an insistence that const values have an initialiser (being static and implicitly
initialised to zero isn't sufficient).