Interrupts working
Managed to program the PIC and the local timer. Needed to update some
Mackerel registers. 32 Bit registers where accessed as 16 Bit
registers, which caused improperly aligned memory accesses (in fact,
even properly aligned accessed with 16 Bit access failed)

As we are not running in user-mode yet, the interrupts are currently
enabled somewhere in the middle of the kernel boot-up. The kernel will
just spin after that and wait for the interrupt. Then, it will crash,
since it does not know how to handle interrupts from the local timer.

Some serious cleanup of the code is required, though.