Steps towards going to user-level.
The first sys_print works (from within the dispatcher of init), but
the state does not seem to be restored properly. The infinite-loop
after this sys_print is never reached. I assume the code to restore
the registers in swi_done is broken.

This code is not stable. A lot of things are disabled, e.g. caching,
interrupts. Also, there are some infinite loops in there.