scc: disable unused-but-set-variable warning (-Wno-unused-but-set-variable)
Notes:
unused-but-set-variable is disabled on most architectures (i.e., x86_{32,64},
ARMv{5,7}). The reason is that it exists only on (at this time) recent gcc
versions and we have code that triggers it. I think it would be a good idea to
enable it so that we can improve the code, but the problem is that flounder
(and probably other C generating tools) generate code subject to that warning.

I've consider the following options:
 1. Fix flounder (and probably other tools) to avoid using uninitialized
    variables. This does not seem to be easy, and it will probably lead to
    duplicated code and make flounder more complicated.
 2. Have flounder add __attribute__((unused)) qualifiers for variables that may
    trigger this warning. This seems too intrusive for flounder -- flounder should
    just care about correctness not silencing compiler warnings.
 3. Disable the warning only for flounder (or other auto-generated) code. I
    think this is the best solution, but I couldn't find a trivial way to
    instruct hake to have additional cflags for flounder generated files.

So, for now I add -Wno-unused-but-set-variable to SCC.