Build fixes for ARMv7.
Make choice about GCC binary. Fixed typos in usbconf mackerel file.
Remove MMU from kernel Hakefile because it is not needed.
Don't do stack protection (leads to compiler errors on newer versions).
Don't add build-id when linking on ARMv7. This screwed up ELF loading in our
kernel when it tried to load a section at 0x8000 -- which is not a valid
virtual address -- for the init binary. For some reason
stripping the section manually did not help.