Enable unaligned memory access on ARMv7.
Some ARM platform are sensitive to unaligned memory access.

To parse the FAT boot-sector we read certain two-byte values
at unaligned offsets. The underlying Mackerel code uses a
uint16_t pointer to do that, which results in a fault.

On ARMv7 we can just enable unaligned memory access in CP15.
Another way around this would be to declare the pointer as
__packed__. However, this is only supported by ARMs official
compiler suite.