kernel: Fixed paging_map_memory() in ARM-based architectures to work for all arguments.
The problem was that for some reason paging_map_memory() was rounding down the
size (in bytes) of the region to map to the next multiple of SECTION_BYTES and
then wouldn't map the last section whenever the value of bytes was a multiple
of SECTION_BYTES already.

Thanks to Jeremia Bär and Ralf Sager (AOS students) for alerting us to the
problem.