Increase check for module_end from 4MB to 4GB.
According to the old hg repo the 4 MB limit was introduced at a point where we
only identity mapped the lower 4 MB of the address space. Now all of the 4GB
address space is identity mapped in the bootstrap code.

Changing this limit allows for the bootstrap CPU driver to load the kernel
modules conveniently from vfs instead of relying on multiboot to get the
module (but now the frame we allocate for the binary is in a higher address).