Add syscall to get phys address of struct global.
We need to set a pointer in the bootstrap code of the new kernel to
point to this global struct that is shared by all kernels. The global
struct contains a lock for printing to console.

This change allows to completely remove all boot-up logic in the kernel.
All we use the kernel for now is to send interrupts to awake the
destination core.