Construct a kcb in x86boot.
Pass it along with the rest of coredata in the kernel.
We currently push the kcb in octopus to make sure have a reference to the cap
during multiple invocations of x86boot. Maybe it would be cleaner if kaluga
constructs the kcb caps and passes it on to x86boot.