cp15: use clean & invalidate instead of clean only
The current implementation of cp15_invalidate_d_cache() only cleans
the cache, which means that dirty cache lines are written back to
memory. However, cache lines which were modified in memory, for
example by a DMA engine, are not re-read.

This change uses the clean & invalidate operation, which means that
the cache lines are cleaned and invalidated: All dirty cache lines
are written back to memory, all other other cache lines are marked
as invalid until they are re-read from memory.