Addition of a second supercn.
It turned out that running Barrelfish on machines with lots of RAM will
results in a error during booting that the supercn has no more space
left. As soon as supercn0 has no space left, the allocator will switch
to the second supercn1 to store the capabilities.

Also memserver will get both endpoints and switches automatically to the
next supercn while building up the memory allocator. The maximum size
bits are extended to 40.

The offsets are adapted to support 1024GB physical memory.

The implementation has been tested on qemu prior commit.