Make vmkit test-case pass again.
vmkit linux execution stopped after:
[0:05:20] handle_vmexit_cpuid: CPUID: func 40000000, host reports: eax 0, ebx 0,
ecx 0, edx 0
[0:05:20] VMKit PCI host: reading register 2, opsize 1: 0
[0:05:20] VMKit PCI host: reading register 0, opsize 1: 0
[0:05:20] VMKit PCI host: reading register 0, opsize 1: 0
[0:05:20] VMKit PCI host: reading register 3, opsize 0: 0
[0:05:20] handle_vmexit_cpuid: CPUID: func 40000000, host reports: eax 0, ebx 0,
ecx 0, edx 0
[0:16:13] [Error: test timed out]

There seems to be a hard-coded assumption in there about the
PCI function of the NIC it uses for PCI pass-through?
I did not find it with a quick glance, but I'll need to have
another look as this will likely break again in the future.