PCI: Fixed PCI bus configuration on nos6
The problem was introduced with the Arrakis changes. From Simon
Peter's description:

The new PCI configuration will reassign all discovered buses to
space them 2 bus numbers apart, so that SR-IOV devices can stride into
the gaps.

nos6 does not seem to support the full range of 256 bus members. I
added a check to only execute the new code for devices that are
assigned to bus numbers supported by the machine.