Bugfixes in Xeon Phi DMA implementation:
Client: When the done notification was sent right after the RPC reply
the done notification may be handled before the RPC reply causing an
unknown transfer ID exception.

Xeon Phi DMA: Bug in the calculation of the number of descriptors to
process caused a bug with the transfer ordering.

Xeon Phi DMA: changed the processing of the descriptors as the appear to
be done and not waiting untill the entire transfer is done.

Descriptor rings: enforcing to have a power of two size.

DMA Benchmark: Added DMA benchmark to the xeon phi driver.