DMA library: Bugfixes and missing implementations for IOAT
Moved the missing parts of the IOAT DMA driver into the library and
fixed some bugs in the initialization. Basic NOP descriptor test worked.

Also changed the channel get status queries to not accessing the device
register by default.

Changed the completion address reading to the memory region instead of
the register.

Added a new interface for IOAT DMA manager domain. (the driver domain)

Renaimed the scope of the error codes to DMA_ERR_*