libbomp: merged xomp and bomp together and refactoration
There may be a use case where the bomp and xomp library are used
together i.e. we have a linker conflict and not the proper functions are
called.

Refactoration: instead of global variables there is a single state
pointer which can be exchanged depending on the backend to be used.
Further function pointers in this state ensure that the correct
functions are called.

added more sub functions for GCC pragma function calls.

moved omp.h into the global include directory.

merged: bomp_init with bomp_span_domain