Cleaning up the fish mess.
Not sure why we had two files fish.c and fish_arm.c with the same content
except for the reboot and poweroff functions.
Now we have only one again. Next step is to provide a platform independent
abstraction for reboot and poweroff in the OS since we don't really want platform
specific shells...
Also added nproc command that prints the should print amount of OSnodes
in the system.