Xeon Phi boot: getting rid of the single boot image and dependencies
Adapting the boot protocol of the Xeon Phi cards to dynamically
load the individual modules at run time rather than compile time.

The following sources can be used to load the modules
- in ramfs of host machine
- from a TFTP server (using the --tftp flag)
- from NFS (using the --nfs flag)

The modules to be loaded are specified in a supplied menu.lst file.
The parameter  --modlist can be used to specify that file.

The boot driver will handle the generation of the needed data structures
and headers when the card boots.

Signed-off-by: Reto Achermann <reto.achermann@inf.ethz.ch>