Made errval_t a flounder builtin.
The errval type is used internally by flounder, but has since been duplicated
to be passed around through interfaces by user code.  This led to a
proliferation of ifname_errval_t types, which were used interchangably in
user-level code.  This breaks once strict typedef checking is enforced, as in
recent GCC versions.  This is a bandaid fix, and there's a larger underlying
issue still to be resolved around how to deal with globally-defined types in
flounder.

Signed-off-by: David Cock <david.cock@inf.ethz.ch>