kernel: do not drop RAM caps
When deleting last copy of a RAM-derived cap without going through the monitor
it is possible that the monitor endpoint is already occupied.  If that's the
case we used to just leak the reclaimed RAM cap.  This patch changes the
behaviour in that case to return SYS_ERR_RETRY_THROUGH_MONITOR which then
provides a free slot to put the freshly created RAM cap.

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>