pci: Expose real bar number to pci clients.
In Barrelfish, the PCI BARs are compacted into a non-sparse array. Some devices
have bars that have gaps between them, for example 0 and 2, or 0 and 5. This
information can be inportant to drivers. In order to address this, this change
returns the bar number on the get_cap RPC call and stores it in the device_mem
struct.

Signed-off-by: Moritz Hoffmann <moritz.hoffmann@inf.ethz.ch>