hamlet, kpi types, caps: add pasid type
The type represents a physical address space identifier. This patch adds support
in hamlet for a `pasid' type and declares the corresponding pasid_t and macros
in barrelfish_kpi/types.h. It changes caps.hl to use the new type instead of
uint32. The pasid type is represented by a uint32 at the moment.

Signed-off-by: Moritz Hoffmann <moritz.hoffmann@inf.ethz.ch>