Bug fix for thc_async
affected files: barrelfish/include/thc/thcinternal.h
Issue: When the ASYNC macro (in include/thc/thc.h) is called
many times on an x86_64 machine, the stack overflows.

Fix: The stack needs to be popped once more in the RETURN_CONT
macro (in include/thc/thcinternal.h).

In particular, when ASYNC is called many times without there
being a THCYield that happens in the ASYNC block, the stack
overflows. This is because a jump instruction is used instead
of a ret instruction, so in addition to popping the stack once
to get the old ebp value, it must be popped again since space
was reserved for the old return address. This bug causes the stack
to increase by eight bytes after every time ASYNC is called without
a THCYield being invoked from within it.

Signed-off-by: Michael F Quigley <michaelForrQuigley@gmail.com>