T191: Implement prototype for new retype
The new retype call takes an offset into the source cap and a count of objects
to create.  This is different from the old call that just took a source and
created as many new objects as fit into that source (determined by source size
and requested size for new objects).

Currently available as `cap_retype2(dest_start, src, offset, new_type, objsize, count);`

Note that this is not used anywhere yet.

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>