harness,qemu wrapper: Make harness use qemu wrapper script for x86_64
wrapper script understands new option --smp to specify the number of cores for
qemu. wrapper script also looks for 'root ...' line in menu.lst and prepends
the argument to the kernel and initrd files when it's not '(nd)'.

harness writes a sensible root statement into its temporary qemu menu.lst.

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>