Rationalised the ARM Platform code.
We now have separate targets for the GEM5 VE and FVP VE, as the GEM5 version
actually simulates an A15, and has a few quirks.  The FVP VE and Pandaboard
targets share most of their code.  Thus we've got three ARMv7-A targets now,
and it should be easy to add more, as soon as we've got the simulators.

Signed-off-by: David Cock <david.cock@inf.ethz.ch>