Rewrote do_resume() for ARMv7.
The previous code was trying to return to user mode (USR), directly from
system mode.  You can't do that, at least not the way we were trying to, as
system mode has no SPSR (that's why GEM5 has been complaining about writes to
SPSR from an invalid mode for the last half a decade), and no banked
registers.  In any case, LDMIA {PC}^ is UNDEFINED from system mode - therefore
the ARMv7 dispatch code has always been functioning as a complete fluke!

I intend to rewrite this again to use RFE, which is the only supported way to
return to user level without chaining through an exception mode, but this fix
is enough to get it running, and squash that warning at last.

n.b. the top two hits for "Trying to access SPSR in an invalid mode" are
people trying to run Barrelfish!

Signed-off-by: David Cock <david.cock@inf.ethz.ch>