ARMv7: Rationalised coreboot structures.
Put the Multiboot header in the coredata struct, and allocate both in and the
globals (which now contains only the printf spinlock) in the BSP kernel's BSS.
We no longer rely on magic addresses.

Doing so exposed a latent bug, where code using the various alloc_phys()
functions were assuming that the output would be page-aligned, when that
wasn't guaranteed.  Added alloc_phys_aligned to all relevant prototypes.

Signed-off-by: David Cock <david.cock@inf.ethz.ch>