ARMv7: Separated pre-MMU init code
The ARMv7 initialisation is now divided strictly into a pre-MMU phase and a
post-MMU phase, with the jump to arch_init_2 being a calculated long jump.
This doesn't make much difference on the Pandaboard and other platforms with
RAM at 0x80000000, but it makes it possible to support other RAM locations
with little or no change to the code.  This will be necessary for the Zynq.

There's a little assorted cleanup included in this patch, of the MMU init
code.

Signed-off-by: David Cock <david.cock@inf.ethz.ch>