T258: refactor creation of new cspaces to make it cleaner
This removes the need for manual updating of cnoderefs in libspawndomain. We
introduce a new function cnode_create_l1() for creating new L1 cnodes in L2
slots in our cspace, and rewrite cnode_create_foreign() to
cnode_create_foreign_l2() which takes a capref to an L1 cnode and a slot in
the L1 cnode for where to put the new L2 cnode.

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>