ARMv7: Coreboot paramaters through the SKB
For ARMv7 platforms, this patch modifies Kaluga and Coreboot to discover the
available CPUs, and their boot parameters (e.g. CPU driver binary) via the
SKB, where this data is preseeded, as we can't easily probe a lot of this
without ACPI.

This *should* let us implement Coreboot pretty easily for A15 and A17.

Signed-off-by: David Cock <david.cock@inf.ethz.ch>