T308: Improve support for Intel 82574 gbe network cards
It turns out the card is too similiar to give it an extra driver, remove
the e1000e binary and improve e1000n instead. e1000n needs refactoring.
The driver does not hang on hw reset (on the 82574) and the interrupt
handling is correct enough to let irqtest pass (with legacy interrupts).

Signed-off-by: Lukas Humbel <lukas.humbel@inf.ethz.ch>