T310: capops: Replace num_monitors_online() with num_monitors_ready_for_capops()
This commit replaces num_monitors_online() which was implemented as a query to
octopus for the number of online spawnd processes, with a function that counts
all intermon bindings that have the capops_ready flag set. This is what we
really need to know to decide what to do in the distributed capabilities code.

Closes T310.

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>