libbarrelfish: x86_64: pmap: Fix slab refilling using fixed allocator
If the pmap wants to refill it's slab allocator before a connection to the
memory server is established, and it needs more than a single page for the
refilling -- this is the case where we have array-based shadow page tables in
the pmap -- we need to allocate and map pages slightly differently. This is
now implemented as an extra function which is called when we get
LIB_ERR_RAM_ALLOC_MS_CONSTRAINTS from frame_alloc() in refill_slabs().

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>