Add extra layer of queuing above the Flounder UMP one.
The process manager now handles its own high-level queues, over the ones
provided by Flounder for UMP transmission. This is meant so that when many
clients have requests that the process manager needs to forward to the same
spawnd instance, "register_send" does not return an "already registered" error.

The queuing implementation follows the one in usr/monitor/queue.c.

Signed-off-by: Razvan Damachi <razvan.damachi@gmail.com>