libc/sys: implement mmap() for anonymous mappings and implement munmap().
This commit provides a partial implementation of mmap() and munmap() for
Barrelfish embedded in the BSD libc. Currently the implementation only
supports creating anonymous mappings (MAP_ANON or MAP_ANONYMOUS). The
MAP_FIXED flag is only supported when MAP_EXCL is specified also. Most other
flags are not implemented.

munmap() supports only unmapping whole existing mappings, and fails when a
caller requests to unmap parts of an existing Barrelfish vregion(). munmap()
does not distinguish between vregions that were created by an mmap() call and
vregions that were created through other means.

There is a new config flag `libc_mmap_nyi_flags` which controls a debug print
statement in mmap() which prints all the specified but not implemented flag
for each call to mmap().

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>