devif: destroy no longer queue specific
To avoid having to know what the queue represents after
having used it, the destroy functionality has to go into
the queue interface. Otherwise, we still have to know
what the queue represents at creation and destruction
time, which does not make sense for destruction since
in between most likely we do not know what exact queue
this is.

Signed-off-by: Roni Häcki <roni.haecki@inf.ethz.ch>