libbarrelfish: pmap: move initial slab buffers back into struct pmap_x86.
When moving the slab buffers into the data section, the redis domain did not
get into user space, but without producing any debuggable error messages.

To mitigate the issue, we move the initial slab buffers back into the pmap_x86
struct. This in turn means that we have a very limited maximum size of 32
initial slabs, as a domain's pmap_x86 struct is contained in the domain's DCB.

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>