kernel: capabilities: mapping types: remove offset field and switch to ptable+entry
To eliminate the need for looking up page table capabilities by address when
cleaning up mappings etc, we switch the mapping capability types to hold a
pointer to the page table capability and the entry in the capability instead
of just a pointer to the first page table entry which corresponds to the
mapping.

To make room for the additional two bytes, we remove the offset into the
mapped capability which was present in the mappings but never used.

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>