queue_service: rough skeleton for serivce
The queue service should be used as a way that network stacks
of applications can request endpoints of NIC drivers from which
queues can be requested. Each endpoint should only be used
once to request a queue.

Up to now this was not a problem since we could directly give the
endpoint to request a driver to the net_socket_server when it is
started. But to generalise this we need a service as described above.

Signed-off-by: Roni Häcki <roni.haecki@inf.ethz.ch>