kernel,libbarrelfish: implement general purpose direct cap_identify
This is required when an application wishes to map its own page tables read
only, as we previously did not have a sensible general purpose way of figuring
out the type of a given capref without using a debug invocation.

Realistically, direct cap_identify should be modified to redact potentially
sensitive information which we want to withhold from applications.

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>