libbarrelfish: early cn allocator: move state to ram_alloc_state
Each dispatcher is provided with a L2 CNode filled with 16kB RAM
capabilities for the purpose of allocating more L2 CNodes before the
connection with the memory server is established.

The first implementation of this early_cn allocator simply used a static
local variable to keep track of which 16kB RAM caps have been handed out
yet. Naturally, this does not work correctly in combination with domain
spanning.

To fix this allocator in the presence of spanned domains, this commit
moves the "next free" indicator into the ram allocator state (struct
ram_alloc_state).

Signed-off-by: Simon Gerber <simon.gerber@inf.ethz.ch>